ci: split Open VSX publish workflow
